refactor(ircx): reorganize RC extraction data structures

+++ src/operation/iRCX/source/toolkit/logger/Logger.hpp
@@ -122,7 +122,13 @@ class Logger
       std::string::size_type pos = file_name.find_last_of('/') + 1;
       file_name = file_name.substr(pos, file_name.length() - pos);
     }
-    std::string prefix = getString("[RCX ", getTimestamp(), " ", getCompressedBase62(std::stoul(getString(std::this_thread::get_id()))), " ", file_name, " ");
+    std::string prefix = getString("[RCX ",
+                                   getTimestamp(),
+                                   " ",
+                                   getCompressedBase62(static_cast<size_t>(std::stoul(getString(std::this_thread::get_id())))),
+                                   " ",
+                                   file_name,
+                                   " ");
     std::string suffix = getString(" ", location.function_name());
     std::string message = getString(value, args...);
 
@@ -181,7 +187,7 @@ class Logger
     return timestamp;
   }
 
-  std::string getCompressedBase62(uint64_t origin)
+  std::string getCompressedBase62(size_t origin)
   {
     std::string base = "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
